pyproject.toml
[project]
name = "monitoring"
version = "0.1.0"
description = "Monitoring framework to test UAS Service Suppliers"
readme = "README.md"
requires-python = "==3.13.5"
dependencies = [
    "aiohttp",
    "arrow",
    "basedpyright>=1.31.1",
    "bc-jsonpath-ng",
    "cryptography",
    "deprecation",
    "faker",
    "flask",
    "flask-login",
    "geojson",
    "gevent",
    "google-auth",
    "graphviz",
    "gunicorn",
    "implicitdict",
    "jsonnet",
    "jsonschema",
    "jwcrypto",
    "kubernetes",
    "locust",
    "loguru",
    "lxml",
    "marko",
    "numpy",
    "oauthlib",
    "pem",
    "pip-tools",
    "psycopg[binary]",
    "pvlib",
    "pyjwt",
    "pykml",
    "pyopenssl",
    "pyproj",
    "pytest",
    "pytest-mock",
    "pyyaml",
    "requests",
    "ruff",
    "s2sphere",
    "scipy",
    "shapely",
    "structlog",
    "termcolor",
    "testcontainers",
    "uas-standards",
    "uuid6",
]

[tool.ruff]
target-version = "py313"

# Default + isort + pyupgrade
lint.select = [
    "E4", "E7", "E9", "F", "I", "UP"
]
extend-exclude = [
  "interfaces/*",
  "monitoring/prober/output/*",
  "monitoring/mock_uss/output/*",
  "monitoring/uss_qualifier/output/*",
]
line-length = 88

[tool.basedpyright]
typeCheckingMode = "standard"
failOnWarnings = true

exclude = [
  "**/__pycache__",
  "**/.*",
  "interfaces/*",
  "monitoring/prober/output/*",
  "monitoring/mock_uss/output/*",
  "monitoring/uss_qualifier/output/*",
]

[dependency-groups]
dev = [
    "types-lxml>=2025.8.25",
]
